Further refinement of table driven code
Pull more information from "data" sheets in the spreadsheet so that over time the behavior of the spreadsheet can be changed just by changing the data page, and no need to edit the VBA.
Items like the temp directory, the spreadsheet name, are no longer hardcoded. Also items like the filter area are now pulled from a data tab.